RulesDataManagement/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
twilio==7.16.5
waitress==2.1.2
